requirements.txt
flask
firebase
pyrebase4
firebase_admin
langchain
PyPDF2
langchain_google_genai 
generativeai
python-dotenv
supabase
faiss
faiss-gpu
